Added vector wrapper; Added multithreaded model loading using assetRegistry; Added texture support to materials; Added xxhash; Added fallback textures; Added uniqueRef array support; Added thread priority support;  Added shader files to premake; Added shadows; Added asset icons; Added asset browser filtering & imguiSelection; Added first version of scene serialization; Improved context based job system with priority and grouping; Improved textureLoader; Improved material uniform uploading; Improved texture loading; Improved textureDesc; Improved color class; Improved UUID generation; Updated Premake5.

+++ engine/dependencies/imguiProfiler/ProfilerWindow.cpp
@@ -410,7 +410,7 @@ static void DrawProfilerTimeline(const ImVec2& size = ImVec2(0, 0)) {
 					context.IsSelectingRange = false;
 				} else {
 					// Distance between mouse cursor and measuring start
-					float distance = fabs(ImGui::GetMousePos().x - context.RangeSelectionStart);
+					float distance = (float)fabs(ImGui::GetMousePos().x - context.RangeSelectionStart);
 
 					// Fade in based on distance
 					float opacity = ImClamp(distance / 30.0f, 0.0f, 1.0f);
